requirements-dev.txt
ruff==0.9.3
pytest==8.3.3
pytest-cov==5.0.0
wheel==0.44.0
build==1.2.2.post1

mysql-connector-python==9.1.0
numpy==2.1.2
python-slugify==8.0.4
pytz==2024.2
